Async cluster: move routing stuff to separate file.
