+++ redis/src/cluster_async/mod.rs
@@ -30,13 +30,14 @@ use std::{
     time::Duration,
 };
 
+mod routing;
 use crate::{
     aio::{ConnectionLike, MultiplexedConnection},
     cluster::{get_connection_info, slot_cmd},
     cluster_client::{ClusterParams, RetryParams},
     cluster_routing::{
-        self, MultipleNodeRoutingInfo, Redirect, ResponsePolicy, Route, RoutingInfo,
-        SingleNodeRoutingInfo, Slot, SlotAddr, SlotMap,
+        MultipleNodeRoutingInfo, Redirect, ResponsePolicy, RoutingInfo, SingleNodeRoutingInfo,
+        Slot, SlotMap,
     },
     cluster_topology::parse_slots,
     Cmd, ConnectionInfo, ErrorKind, IntoConnectionInfo, RedisError, RedisFuture, RedisResult,
@@ -49,6 +50,7 @@ use futures::{future::BoxFuture, prelude::*, ready};
 use log::{trace, warn};
 use pin_project_lite::pin_project;
 use rand::{seq::IteratorRandom, thread_rng};
+use routing::{route_for_pipeline, InternalRoutingInfo, InternalSingleNodeRouting};
 use tokio::sync::{mpsc, oneshot, RwLock};
 
 /// This represents an async Redis Cluster connection. It stores the
@@ -170,66 +172,6 @@ struct ClusterConnInner<C> {
 }
 
 #[derive(Clone)]
-enum InternalRoutingInfo<C> {
-    SingleNode(InternalSingleNodeRouting<C>),
-    MultiNode((MultipleNodeRoutingInfo, Option<ResponsePolicy>)),
-}
-
-impl<C> From<cluster_routing::RoutingInfo> for InternalRoutingInfo<C> {
-    fn from(value: cluster_routing::RoutingInfo) -> Self {
-        match value {
-            cluster_routing::RoutingInfo::SingleNode(route) => {
-                InternalRoutingInfo::SingleNode(route.into())
-            }
-            cluster_routing::RoutingInfo::MultiNode(routes) => {
-                InternalRoutingInfo::MultiNode(routes)
-            }
-        }
-    }
-}
-
-impl<C> From<InternalSingleNodeRouting<C>> for InternalRoutingInfo<C> {
-    fn from(value: InternalSingleNodeRouting<C>) -> Self {
-        InternalRoutingInfo::SingleNode(value)
-    }
-}
-
-#[derive(Clone)]
-enum InternalSingleNodeRouting<C> {
-    Random,
-    SpecificNode(Route),
-    ByAddress(String),
-    Connection {
-        identifier: String,
-        conn: ConnectionFuture<C>,
-    },
-    Redirect {
-        redirect: Redirect,
-        previous_routing: Box<InternalSingleNodeRouting<C>>,
-    },
-}
-
-impl<C> Default for InternalSingleNodeRouting<C> {
-    fn default() -> Self {
-        Self::Random
-    }
-}
-
-impl<C> From<SingleNodeRoutingInfo> for InternalSingleNodeRouting<C> {
-    fn from(value: SingleNodeRoutingInfo) -> Self {
-        match value {
-            SingleNodeRoutingInfo::Random => InternalSingleNodeRouting::Random,
-            SingleNodeRoutingInfo::SpecificNode(route) => {
-                InternalSingleNodeRouting::SpecificNode(route)
-            }
-            SingleNodeRoutingInfo::ByAddress { host, port } => {
-                InternalSingleNodeRouting::ByAddress(format!("{host}:{port}"))
-            }
-        }
-    }
-}
-
-#[derive(Clone)]
 enum CmdArg<C> {
     Cmd {
         cmd: Arc<Cmd>,
@@ -241,41 +183,6 @@ enum CmdArg<C> {
         count: usize,
         route: InternalSingleNodeRouting<C>,
     },
-}
-
-fn route_for_pipeline(pipeline: &crate::Pipeline) -> RedisResult<Option<Route>> {
-    fn route_for_command(cmd: &Cmd) -> Option<Route> {
-        match cluster_routing::RoutingInfo::for_routable(cmd) {
-            Some(cluster_routing::RoutingInfo::SingleNode(SingleNodeRoutingInfo::Random)) => None,
-            Some(cluster_routing::RoutingInfo::SingleNode(
-                SingleNodeRoutingInfo::SpecificNode(route),
-            )) => Some(route),
-            Some(cluster_routing::RoutingInfo::MultiNode(_)) => None,
-            Some(cluster_routing::RoutingInfo::SingleNode(SingleNodeRoutingInfo::ByAddress {
-                ..
-            })) => None,
-            None => None,
-        }
-    }
-
-    // Find first specific slot and send to it. There's no need to check If later commands
-    // should be routed to a different slot, since the server will return an error indicating this.
-    pipeline.cmd_iter().map(route_for_command).try_fold(
-        None,
-        |chosen_route, next_cmd_route| match (chosen_route, next_cmd_route) {
-            (None, _) => Ok(next_cmd_route),
-            (_, None) => Ok(chosen_route),
-            (Some(chosen_route), Some(next_cmd_route)) => {
-                if chosen_route.slot() != next_cmd_route.slot() {
-                    Err((ErrorKind::CrossSlot, "Received crossed slots in pipeline").into())
-                } else if chosen_route.slot_addr() != &SlotAddr::Master {
-                    Ok(Some(next_cmd_route))
-                } else {
-                    Ok(Some(chosen_route))
-                }
-            }
-        },
-    )
 }
 
 fn boxed_sleep(duration: Duration) -> BoxFuture<'static, ()> {
@@ -1472,88 +1379,4 @@ where
                 .get(addr)
                 .map(|conn| (addr.clone(), conn.clone()))
         })
-}
-
-#[cfg(test)]
-mod pipeline_routing_tests {
-    use super::route_for_pipeline;
-    use crate::{
-        cluster_routing::{Route, SlotAddr},
-        cmd,
-    };
-
-    #[test]
-    fn test_first_route_is_found() {
-        let mut pipeline = crate::Pipeline::new();
-
-        pipeline
-            .add_command(cmd("FLUSHALL")) // route to all masters
-            .get("foo") // route to slot 12182
-            .add_command(cmd("EVAL")); // route randomly
-
-        assert_eq!(
-            route_for_pipeline(&pipeline),
-            Ok(Some(Route::new(12182, SlotAddr::ReplicaOptional)))
-        );
-    }
-
-    #[test]
-    fn test_return_none_if_no_route_is_found() {
-        let mut pipeline = crate::Pipeline::new();
-
-        pipeline
-            .add_command(cmd("FLUSHALL")) // route to all masters
-            .add_command(cmd("EVAL")); // route randomly
-
-        assert_eq!(route_for_pipeline(&pipeline), Ok(None));
-    }
-
-    #[test]
-    fn test_prefer_primary_route_over_replica() {
-        let mut pipeline = crate::Pipeline::new();
-
-        pipeline
-            .get("foo") // route to replica of slot 12182
-            .add_command(cmd("FLUSHALL")) // route to all masters
-            .add_command(cmd("EVAL"))// route randomly
-            .cmd("CONFIG").arg("GET").arg("timeout") // unkeyed command
-            .set("foo", "bar"); // route to primary of slot 12182
-
-        assert_eq!(
-            route_for_pipeline(&pipeline),
-            Ok(Some(Route::new(12182, SlotAddr::Master)))
-        );
-    }
-
-    #[test]
-    fn test_raise_cross_slot_error_on_conflicting_slots() {
-        let mut pipeline = crate::Pipeline::new();
-
-        pipeline
-            .add_command(cmd("FLUSHALL")) // route to all masters
-            .set("baz", "bar") // route to slot 4813
-            .get("foo"); // route to slot 12182
-
-        assert_eq!(
-            route_for_pipeline(&pipeline).unwrap_err().kind(),
-            crate::ErrorKind::CrossSlot
-        );
-    }
-
-    #[test]
-    fn unkeyed_commands_dont_affect_route() {
-        let mut pipeline = crate::Pipeline::new();
-
-        pipeline
-            .set("{foo}bar", "baz") // route to primary of slot 12182
-            .cmd("CONFIG").arg("GET").arg("timeout") // unkeyed command
-            .set("foo", "bar") // route to primary of slot 12182
-            .cmd("DEBUG").arg("PAUSE").arg("100") // unkeyed command
-            .cmd("ECHO").arg("hello world"); // unkeyed command
-
-        assert_eq!(
-            route_for_pipeline(&pipeline),
-            Ok(Some(Route::new(12182, SlotAddr::Master)))
-        );
-    }
 }
